+++ components/user-mgt/pom.xml
@@ -38,6 +38,7 @@
         <module>org.wso2.carbon.user.mgt</module>
         <module>org.wso2.carbon.user.mgt.common</module>
         <module>org.wso2.carbon.user.mgt.ui</module>
+        <module>org.wso2.carbon.user.mgt.workflow</module>
         <module>org.wso2.carbon.identity.user.registration</module>
         <module>org.wso2.carbon.identity.user.profile</module>
         <module>org.wso2.carbon.identity.user.profile.ui</module>
