+++ components/workflow-mgt/pom.xml
@@ -36,7 +36,6 @@
 
     <modules>
         <module>org.wso2.carbon.identity.workflow.mgt</module>
-        <module>org.wso2.carbon.identity.workflow.mgt.impl</module>
         <module>org.wso2.carbon.identity.workflow.mgt.ui</module>
     </modules>
