+++ features/user-mgt-workflow/org.wso2.carbon.user.mgt.workflow.server.feature/pom.xml
@@ -21,17 +21,18 @@
 
     <parent>
         <groupId>org.wso2.carbon.identity</groupId>
-        <artifactId>workflow-mgt-feature</artifactId>
+        <artifactId>user-mgt-workflow-feature</artifactId>
         <version>4.5.4-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>org.wso2.carbon.identity.workflow.mgt.server.feature</artifactId>
+    <artifactId>org.wso2.carbon.user.mgt.workflow.server.feature</artifactId>
     <packaging>pom</packaging>
-    <name>WSO2 Carbon - Workflow Management Core Feature</name>
+    <name>WSO2 Carbon - User Management Workflow Core Feature</name>
     <url>http://wso2.org</url>
-    <description>This feature contains the core bundles required for Back-end Workflow Management functionality
+    <description>This feature contains the core bundles required for Back-end User Management Workflow
+        functionality
     </description>
 
     <dependencies>
@@ -39,12 +40,7 @@
             <groupId>org.wso2.carbon.identity</groupId>
             <artifactId>org.wso2.carbon.identity.workflow.mgt</artifactId>
         </dependency>
-        <dependency>
-            <groupId>org.wso2.carbon.identity</groupId>
-            <artifactId>org.wso2.carbon.identity.workflow.mgt.bps.stub</artifactId>
-        </dependency>
     </dependencies>
-
     <build>
         <plugins>
             <plugin>
@@ -59,7 +55,7 @@
                             <goal>p2-feature-gen</goal>
                         </goals>
                         <configuration>
-                            <id>org.wso2.carbon.identity.workflow.mgt.server</id>
+                            <id>org.wso2.carbon.user.mgt.workflow.server</id>
                             <propertiesFile>../../etc/feature.properties</propertiesFile>
                             <adviceFile>
                                 <properties>
@@ -68,11 +64,10 @@
                                 </properties>
                             </adviceFile>
                             <bundles>
-                                <bundleDef>org.wso2.carbon.identity:org.wso2.carbon.identity.workflow.mgt.bps.stub</bundleDef>
-                                <bundleDef>org.wso2.carbon.identity:org.wso2.carbon.identity.workflow.mgt</bundleDef>
+                                <bundleDef>org.wso2.carbon.identity:org.wso2.carbon.user.mgt.workflow</bundleDef>
                             </bundles>
                             <importFeatures>
-                                <importFeatureDef>org.wso2.carbon.identity.core.server</importFeatureDef>
+                                <importFeatureDef>org.wso2.carbon.identity.workflow.mgt:${carbon.identity.version}</importFeatureDef>
                             </importFeatures>
                         </configuration>
                     </execution>
